fix: coder eval for salesforce (#ENGCE-58686)
