+++ tests/tasks/uipath-maestro-flow/connector_features/generic_dynamic_node/generic_dynamic_node.yaml
@@ -1,0 +1,77 @@
+task_id: skill-flow-generic-dynamic-node
+description: >
+  Connector feature: validate a generic (dynamic) connector node end-to-end. A
+  generic activity encodes only the operation in its node type; the object is
+  supplied dynamically at configure time, so the agent must resolve the object
+  name and set it on the node. Uses ServiceNow `list-all-records`
+  (`uipath.connector.uipath-servicenow-servicenow.list-all-records`, API
+  `objectName: "acr_user"`) as the concrete generic activity, bound to the
+  tenant's ServiceNow connection, then runs `flow debug`. Exercises connector
+  discovery (including connections in non-default folders), generic-activity
+  object-name resolution, connection binding, and live execution. The `acr_user`
+  table is empty in the codereval tenant, so the call legitimately returns `[]`;
+  the checker asserts the node is a generic `list-all-records` node on
+  `objectName: "acr_user"`, `finalStatus: "Completed"`, and an array-typed
+  output (empty allowed) rather than requiring rows.
+  Connection note: ServiceNow developer instances hibernate after inactivity.
+  If a run fails with a connection error, log into the ServiceNow account to
+  wake the instance, then re-run.
+tags: [uipath-maestro-flow, e2e, mode:operate, lifecycle:generate, shape:single-node, connector, uipath-servicenow-servicenow, ipe]
+
+agent:
+  type: claude-code
+  permission_mode: acceptEdits
+  allowed_tools: ["Skill", "Bash", "Read", "Write", "Edit", "Glob", "Grep"]
+
+run_limits:
+  task_timeout: 1500
+  max_turns: 120
+  turn_timeout: 1200
+
+initial_prompt: |
+  Create a UiPath Flow project named "AcrUserList" with a manual trigger.
+  The process must call ServiceNow to list all records in the Acr User
+  object, then surface the returned records as a flow output variable.
+
+  Bind it to the tenant's ServiceNow connection. Then run the flow with
+  `uip maestro flow debug` to confirm the connector call completes.
+
+  Use the connection present in Shared/uipath-maestro-flow.
+
+  Do NOT ask for approval, confirmation, or feedback. Do NOT pause between
+  planning and implementation. Build, validate, and run end-to-end in a single
+  pass. Before starting, load the uipath-maestro-flow skill and follow its
+  connector node configuration workflow. Do NOT substitute a mock for the
+  connector.
+
+success_criteria:
+  # ── Flow file validity ─────────────────────────────────────────────────
+  - type: run_command
+    description: "uip maestro flow validate passes on the flow file"
+    command: "uip maestro flow validate AcrUserList/AcrUserList/AcrUserList.flow"
+    timeout: 30
+    expected_exit_code: 0
+    weight: 3.0
+    pass_threshold: 1.0
+
+  # ── Agent invoked debug ────────────────────────────────────────────────
+  - type: command_executed
+    description: "Agent ran flow debug with --output json"
+    tool_name: "Bash"
+    command_pattern: '(uip|\$UIP|\$\{?UIP\}?)\s+(maestro\s+)?flow\s+debug.*--output\s+json'
+    min_count: 1
+    weight: 1.5
+    pass_threshold: 1.0
+
+  # ── Execution: connector calls ServiceNow and surfaces an array output ──
+  - type: run_command
+    description: "ServiceNow list-all-records node on acr_user executes and surfaces an array output"
+    command: "python3 $TASK_DIR/check_generic_dynamic_node.py"
+    timeout: 320
+    expected_exit_code: 0
+    weight: 6.0
+    pass_threshold: 1.0
+
+post_run:
+  - command: "python3 $SKILLS_REPO_PATH/tests/tasks/uipath-maestro-flow/_shared/cleanup_solutions.py"
+    timeout: 120
